adding character_value relations
separating related and non-related characters

listing character relations in value edit

delete button exists

switch alignment is now a button

deleting and reloading

invert alignments

Editing relations

getting to the "create value change" screen

loading values to change

TO DO comment

major style changes

more style stuff

value change exists now

updated readme

added serif font

showing value changes, plus font update

fix function

loading value_change for edit

update value change function

creating and updating value_changes freely

removed unused functions

notification exists and works

improved notifications

notifications work better in all cases

to do added

loading to previous spot

load x component buttons created

show btns on story level

new value change returns now

return links robust

fixed bug

removed button at end of list

showing the beats

loading beats better

sorting components

some testing functions

comment

removed buttons from showAll cards

go to level to edit

fixed some BUGS

show all components is DONE

new components' labels are numbered now

value objects pageTitle update after create

changed story.html to cards.html

listing all value changes

required fields enforced

updating value objects checks value first

deleting works

deleting with notification

edit story page styled

subcomponent editing page styled

edit story notifications

"edit value change" navigation

starting the "archive story" logic

archiving stories

deleting is progressing

progress AND errors

+++ public/js/html.js
@@ -1,7 +1,11 @@
 import * as helpers from 'helpers'
+import * as consts from 'consts'
+
+// Construct & return name of function to load component
+const loadFunctionName = (level, componentId) => "loadComponent(\"" + level + "\", " + String(componentId) + ")"
 
 const elements = {
-    // STORY.HTML
+    // CARDS.HTML
 
     gridX: () => {
         const div = document.createElement("div")
@@ -9,7 +13,8 @@ const elements = {
         return div
     },
 
-    card: (storyComponent, level, levels) => {
+    // Story components (acts, chapters, scenes, beats) are listed on these cards.
+    card: (storyComponent, level, isLast, skipNav) => {
         // create the HTML elements
         const gridCell = document.createElement("div")
         const callout = document.createElement("div")
@@ -18,44 +23,72 @@ const elements = {
         const cardLabelContainer = document.createElement("div")
         const orderPar = document.createElement("h5")
 
-
-        // Get name of loading function (card has link to its story component)
-        const loadFunctionName = level == levels.ACT ? "loadAct(" :
-            level == levels.CHAPTER ? "loadChapter(" :
-                level == levels.SCENE ? "loadScene(" :
-                    level == levels.BEAT ? "loadBeat(" : ""
-
-        gridCell.setAttribute("class", "large-4 medium-4 small-6 cell")
+        gridCell.setAttribute("class", "large-3 medium-4 small-6 cell")
         callout.setAttribute("class", "callout")
         cardLabel.setAttribute("class", "cardLabel")
-        cardLabel.setAttribute("onclick", loadFunctionName + storyComponent.id + ")")
+        cardLabel.setAttribute("onclick", loadFunctionName(level, storyComponent.id))
 
         cardLabelContainer.setAttribute("class", "cardLabelContainer")
         cardLabelContainer.appendChild(cardLabel)
 
-        // create the "change order" buttons
+        // create this now so we can use it after the 'if' block.
         const bottomMenu = document.createElement("div")
-        const newComponentButtonLeft = document.createElement("div")
-        const deleteButtonMiddle = document.createElement("div")
-        const switchButtonRight = document.createElement("div")
 
-        newComponentButtonLeft.setAttribute("class", "newComponentButtonLeft")
-        newComponentButtonLeft.innerHTML = "<< new " + level
-        newComponentButtonLeft.setAttribute("onclick", "newComponentLeft('" + level + "', " + storyComponent.order + ")")
+        if (!skipNav) {
+            // create the "change order" buttons
+            const newComponentButtonLeft = document.createElement("a")
+            const deleteButtonMiddle = document.createElement("a")
+            const switchButtonRight = document.createElement("a")
 
-        deleteButtonMiddle.setAttribute("class", "deleteButtonMiddle")
-        deleteButtonMiddle.setAttribute("onclick", "deleteComponentRequest('" + level + "', " + storyComponent.id + ")")
-        deleteButtonMiddle.innerHTML = "[delete]"
+            newComponentButtonLeft.setAttribute("class", "newComponentButtonLeft componentCardNav")
+            newComponentButtonLeft.innerHTML = "<< new " + level
+            newComponentButtonLeft.setAttribute("onclick", "newComponentLeft('" + level + "', " + storyComponent.order + ")")
+            // For accessibility, make it tabbable and make "Enter" key function as click.
+            newComponentButtonLeft.setAttribute("tabIndex", 0)
+            newComponentButtonLeft.addEventListener("keypress", function (event) {
+                if (event.key === "Enter") {
+                    event.preventDefault();
+                    newComponentButtonLeft.click();
+                }
+            })
 
-        switchButtonRight.setAttribute("class", "switchButtonRight")
-        switchButtonRight.setAttribute("id", "switchButtonRight_" + level + "_order-" + storyComponent.order)
-        switchButtonRight.innerHTML = "move >>"
-        switchButtonRight.setAttribute("onclick", "shiftComponentRight('" + level + "', " + storyComponent.id + ")")
+            deleteButtonMiddle.setAttribute("class", "deleteButtonMiddle componentCardNav")
+            deleteButtonMiddle.setAttribute("onclick", "deleteComponentRequest('" + level + "', " + storyComponent.id + ")")
+            deleteButtonMiddle.innerHTML = "[delete]"
 
-        bottomMenu.setAttribute("class", "cardBottomMenu")
-        bottomMenu.appendChild(newComponentButtonLeft)
-        bottomMenu.appendChild(deleteButtonMiddle)
-        bottomMenu.appendChild(switchButtonRight)
+            // For accessibility, make it tabbable and make "Enter" key function as click.
+            deleteButtonMiddle.setAttribute("tabIndex", 0)
+            deleteButtonMiddle.addEventListener("keypress", function (event) {
+                if (event.key === "Enter") {
+                    event.preventDefault();
+                    deleteButtonMiddle.click();
+                }
+            })
+
+            switchButtonRight.setAttribute("id", "switchButtonRight_" + level + "_order-" + storyComponent.order)
+            switchButtonRight.innerHTML = "move >>"
+
+            // Don't put the "move right" button on the last card.
+            if (!isLast) {
+                switchButtonRight.setAttribute("class", "switchButtonRight componentCardNav")
+                switchButtonRight.setAttribute("onclick", "shiftComponentRight('" + level + "', " + storyComponent.id + ")")
+                // For accessibility, make it tabbable and make "Enter" key function as click.
+                switchButtonRight.setAttribute("tabIndex", 0)
+                switchButtonRight.addEventListener("keypress", function (event) {
+                    if (event.key === "Enter" && !isLast) {
+                        event.preventDefault();
+                        switchButtonRight.click();
+                    }
+                })
+            } else {
+                switchButtonRight.setAttribute("class", "nullButtonRight")
+            }
+
+            bottomMenu.setAttribute("class", "cardBottomMenu")
+            bottomMenu.appendChild(newComponentButtonLeft)
+            bottomMenu.appendChild(deleteButtonMiddle)
+            bottomMenu.appendChild(switchButtonRight)
+        }
 
         orderPar.setAttribute("class", "orderPar")
         orderPar.innerText = storyComponent.order
@@ -66,56 +99,191 @@ const elements = {
         callout.appendChild(orderPar)
         callout.appendChild(cardLabelContainer)
         callout.appendChild(cardDescription)
-        callout.appendChild(bottomMenu)
+        !skipNav && callout.appendChild(bottomMenu)
         gridCell.appendChild(callout)
         return gridCell
     },
 
-    newComponentButton: (level) => {
+    valueChangeCard: (valueChange, storyId, componentChainLinkAddendum) => {
+        const gridCell = document.createElement("div")
+        const callout = document.createElement("div")
+        const link = document.createElement("a")
+        gridCell.setAttribute("class", "large-3 medium-4 small-6 cell")
+        callout.setAttribute("class", "callout")
+        link.setAttribute(
+            "href",
+            "edit_value_object.html?edit=true&value_object_type=value_change&value_object_id=" +
+            valueChange["id"] + "&story_id=" + storyId + componentChainLinkAddendum)
+        link.innerText = valueChange["label"]
+        callout.appendChild(link)
+        gridCell.appendChild(callout)
+        // console.log("ADDENDUM: " + componentChainLinkAddendum)
+        return gridCell
+    },
+
+    newComponentButton: level => {
         // create the HTML elements
         const gridCell = document.createElement("div")
         const callout = document.createElement("div")
-        const cardLabel = document.createElement("h5")
+        const cardLabel = document.createElement("a")
 
         const newComponentFunctionName = "newComponent('" + level + "')"
-        gridCell.setAttribute("class", "large-4 medium-4 small-6 cell")
+        gridCell.setAttribute("class", "large-3 medium-4 small-6 cell")
         callout.setAttribute("class", "callout")
+
         cardLabel.innerText = "Create new " + level
+        cardLabel.setAttribute("tabindex", 0)
+        cardLabel.setAttribute("class", "newElementButton")
         cardLabel.setAttribute("onclick", newComponentFunctionName)
+        cardLabel.addEventListener("keypress", function (event) {
+            if (event.key === "Enter") {
+                event.preventDefault();
+                cardLabel.click();
+            }
+        })
 
         gridCell.appendChild(callout)
         callout.appendChild(cardLabel)
         return gridCell
     },
 
-    valueButton: (storyId, valueId, label) => {
+    newValueChangeButton: (beatId) => {
+        // create the HTML elements
+        const gridCell = document.createElement("div")
+        const callout = document.createElement("div")
+        const cardLabel = document.createElement("a")
+
+        const newValueChangeFunctionName = "newValueChange('" + beatId + "')"
+        gridCell.setAttribute("class", "large-4 medium-4 small-6 cell")
+        callout.setAttribute("class", "callout")
+        cardLabel.innerText = "Create new Value Change"
+        cardLabel.setAttribute("onclick", newValueChangeFunctionName)
+        cardLabel.setAttribute("class", "newElementButton")
+
+        gridCell.appendChild(callout)
+        callout.appendChild(cardLabel)
+        return gridCell
+    },
+
+    valueButton: (storyId, valueId, label, componentChainLinkAddendum) => {
         const button = document.createElement("a")
         button.setAttribute("class", "button small listButton")
         button.innerText = label
         button.setAttribute(
             "href",
-            "edit_value_object.html?edit=true&value_object_type=value&value_object_id=" + valueId + "&story_id=" + storyId)
+            "edit_value_object.html?edit=true&value_object_type=value&value_object_id=" +
+            valueId + "&story_id=" + storyId + componentChainLinkAddendum)
+
         return button
     },
 
-    locationButton: (storyId, locationId, name) => {
+    showAllComponentsButton: (level) => {
+        const onclickFunctionName = "showAllComponents(\"" + level + "\")"
+        const button = document.createElement("a")
+        button.setAttribute("class", "button small whiteButton")
+        button.setAttribute("onclick", onclickFunctionName)
+        button.innerText = level.toUpperCase() + "S"
+        return button
+    },
+
+    locationButton: (storyId, locationId, name, componentChainLinkAddendum) => {
         const button = document.createElement("a")
         button.setAttribute("class", "button small listButton")
         button.innerText = name
         button.setAttribute(
             "href",
-            "edit_value_object.html?edit=true&value_object_type=location&value_object_id=" + locationId + "&story_id=" + storyId)
+            "edit_value_object.html?edit=true&value_object_type=location&value_object_id=" +
+            locationId + "&story_id=" + storyId + componentChainLinkAddendum)
         return button
     },
 
-    characterButton: (storyId, characterId, name) => {
+    characterButton: (storyId, characterId, name, componentChainLinkAddendum) => {
         const button = document.createElement("a")
         button.setAttribute("class", "button small listButton")
         button.innerText = name
         button.setAttribute(
             "href",
-            "edit_value_object.html?edit=true&value_object_type=character&value_object_id=" + characterId + "&story_id=" + storyId)
+            "edit_value_object.html?edit=true&value_object_type=character&value_object_id=" +
+            characterId + "&story_id=" + storyId + componentChainLinkAddendum)
         return button
+    },
+
+    characterValueItemCharacter: (character) => {
+        let charName = character["first_name"]
+        charName += !!character["last_name"] ? " " + character["last_name"] : ""
+
+        const container = document.createElement("div")
+        container.setAttribute("class", "characterValuesListItem")
+
+        const nameDiv = document.createElement("div")
+        nameDiv.setAttribute("class", "characterValuesListItemName")
+
+        const nameText = document.createElement("h6")
+        nameText.innerText = charName
+        nameDiv.appendChild(nameText)
+
+        const alignedDiv = document.createElement("div")
+        alignedDiv.setAttribute("class", "characterValuesListItemAligned")
+        alignedDiv.setAttribute("onclick", "invertAlignment(" + character["character_value"]["id"] + ")")
+        alignedDiv.innerText = !!character["character_value"]["aligned"] ? "aligned" : "against"
+
+        const deleteDiv = document.createElement("div")
+        deleteDiv.setAttribute("class", "characterValuesListItemDelete")
+        deleteDiv.setAttribute("onclick", "deleteCharacterValue(" + character["character_value"]["id"] + ")")
+        deleteDiv.innerText = "DELETE"
+
+        container.appendChild(nameDiv)
+        container.appendChild(alignedDiv)
+        container.appendChild(deleteDiv)
+
+        // console.log("doing a container")
+
+        return container
+    },
+
+    characterValueItemValue: (value) => {
+        const container = document.createElement("div")
+        container.setAttribute("class", "characterValuesListItem")
+
+        const labelDiv = document.createElement("div")
+        labelDiv.setAttribute("class", "characterValuesListItemName")
+
+        const labelText = document.createElement("h6")
+        labelText.innerText = value["label"]
+        labelDiv.appendChild(labelText)
+
+        const alignedDiv = document.createElement("div")
+        alignedDiv.setAttribute("class", "characterValuesListItemAligned")
+        alignedDiv.setAttribute("onclick", "invertAlignment(" + value["character_value"]["id"] + ")")
+        alignedDiv.innerText = !!value["character_value"]["aligned"] ? "aligned" : "against"
+
+        const deleteDiv = document.createElement("div")
+        deleteDiv.setAttribute("class", "characterValuesListItemDelete")
+        deleteDiv.setAttribute("onclick", "deleteCharacterValue(" + value["character_value"]["id"] + ")")
+        deleteDiv.innerText = "DELETE"
+
+        container.appendChild(labelDiv)
+        container.appendChild(alignedDiv)
+        container.appendChild(deleteDiv)
+
+        // console.log("doing a container")
+
+        return container
+    },
+
+    interLevelLabel: (storyComponent, level) => {
+        const cell = document.createElement("div")
+        const label = document.createElement("h3")
+        cell.setAttribute("class", "large-12 cell")
+        label.innerText = storyComponent.label
+
+        if (!!level) {
+            label.setAttribute("onclick", "loadComponent(\"" + level + "\", " + storyComponent.id + ")")
+            label.setAttribute("class", "interLevelLabel")
+        }
+
+        cell.appendChild(label)
+        return cell
     },
 
     // INIT.HTML 
@@ -125,12 +293,26 @@ const elements = {
         return ul
     },
 
-    storyListItem: story => {
+    archiveList: () => {
+        const ul = document.createElement("ul")
+        ul.setAttribute("id", "archiveList")
+        return ul
+    },
+
+    storyListItem: (story, isArchived = false) => {
         const li = document.createElement("li")
         const anchor = document.createElement("a")
         anchor.setAttribute("class", "storiesListItem, button")
-        anchor.setAttribute("href", "story.html?story_id=" + story.id)
+        //anchor.setAttribute("href", "cards.html?story_id=" + story.id)
+        anchor.setAttribute("onclick", "loadStory(" + story.id + ")")
         li.appendChild(anchor)
+        if (!!isArchived) {
+            const deleteButton = document.createElement("a")
+            deleteButton.setAttribute("class", "button small deleteButton")
+            deleteButton.setAttribute("onclick", "requestDelete(" + story.id + ")")
+            deleteButton.innerText = "DELETE"
+            li.appendChild(deleteButton)
+        }
         anchor.innerText = story.label
         return li
     }
